Organize tests for type guessing
